requirements.txt
flask
flask-cors
python-dotenv
groq
pymupdf
youtube-transcript-api
gunicorn
